Add files via upload
Hi all,

I have a question about a problem I am having with the Google Gantt Chart. I am trying to create a chart within google with no dependencies. When put into google, Google re-sorts the dependencies if the dates aren’t corresponding. The results are that the order of resources are rearranged according to the dates. I attached a Word document to this message, were you can see what I mean. The first and above chart is the start situation, the second picture is the resullt of a data change in the recourse: 3.1 Fase 1 (2017). As you can see, the recource 3.1 is moved up in between the numbers 1.2 and 1.3. This is not the desired result. I need the chart to keep the original order of resources and the dependencies to be flexible and not follow corresponding dates. I added this issue to: Google forum/google-visualization-api. 

Unfortunately nobody added an answer or advice. But after own research on the internet, I saw some reactions from other users with simular issues. I have added this link below, and the last person (user “littlecapa”) gives a (possible) answer that seemed to work for him.
 
https://github.com/google/google-visualization-issues/issues/2122
 
Who has a solution or advice? Or can someone confirm the solution that has been given by this user "littlecapa"? I would appreciate this. 
 
Thanks,
 
Regards,
 
Richie